move (image) : move image file
