--- src/main/java/kuit/project/beering/domain/ReviewImage.java
@@ -1,4 +1,0 @@
-package kuit.project.beering.domain;
-
-public class ReviewImage {
-}
